Markdown List Formatting Correction
The list of significant changes wasn't rendering correctly in browsers
due to the absence of a blank line before the start of the list.  I
added it and my editor (Atom) decided to make a ton of other minor
formatting changes (removing whitepsace) to the document as well.  The
only truly useful edit is the addition of line 6.